feat(usability): reseal all at once with centralized functions for scanning
